Carrying capacity scripts output more data
Point-based locations: output isochrones
Polygon-based and point-based locations: output walkable area polygons

+++ scripts/carrying_capacity.py
@@ -4,17 +4,17 @@
 import json
 import geopandas as gp
 from shapely.geometry import shape
-#import area
 import pprint
 import pymongo as db
 import math
-from turfpy.transformation import intersect, circle
+from turfpy.transformation import intersect, circle, difference, union
 from turfpy.measurement import area
-from geojson import (Point, Polygon, Feature, LineString)
-from copy import deepcopy
+from geojson import (Point, Polygon, Feature, LineString, FeatureCollection)
 import sys
 
 pp = pprint.PrettyPrinter(indent=1)
+
+#
 
 if len(sys.argv) < 3:
     print(f"Usage: python {sys.argv[0]} mongodb_url mongodb_database", file=sys.stderr)
@@ -23,42 +23,47 @@ if len(sys.argv) < 3:
 MONGODB_URL = sys.argv[1]
 MONGODB_DATABASE = sys.argv[2]
 
+#
+
 # MONGO CONNECTION INIT
 client = db.MongoClient(MONGODB_URL)
 
+# crowding_db = client["vodafone"]
 crowding_db = client[MONGODB_DATABASE]
 
 osm = crowding_db["osm"]
-osm_no_multi = crowding_db["osm_no_multi"]
 
 # OSM Types collections
 buildings = crowding_db["buildings"]
 water = crowding_db["water"]
 roads = crowding_db["roads"]
+footways = crowding_db["footways"]
 road_polygons = crowding_db["road_polygons"]
 
 sensors = crowding_db["sensors"]
 crowding_polygons = crowding_db["crowding_polygons"]
+usable_areas = crowding_db["usable_areas"]
 
 
 def get_perpendicular_point_1_from_center_at_distance(m, b, center_x, center_y, d):
-    x = (center_x - (m*b) + (center_y * m) +
-         math.sqrt((d**2)*(m**2) - (m**2)*(center_x**2) + (2*center_y*m*center_x) - (2*m*b*center_x) + (d**2) + (2*center_y*b) - (center_y**2) - (b**2)))/\
-        (1+m**2)
-    y = m*x + b
+    x = (center_x - (m * b) + (center_y * m) +
+         math.sqrt((d ** 2) * (m ** 2) - (m ** 2) * (center_x ** 2) + (2 * center_y * m * center_x) - (
+                 2 * m * b * center_x) + (d ** 2) + (2 * center_y * b) - (center_y ** 2) - (b ** 2))) / \
+        (1 + m ** 2)
+    y = m * x + b
     return x, y
 
 
 def get_perpendicular_point_2_from_center_at_distance(m, b, center_x, center_y, d):
-    x = (center_x - (m*b) + (center_y * m) -
-         math.sqrt((d**2)*(m**2) - (m**2)*(center_x**2) + (2*center_y*m*center_x) - (2*m*b*center_x) + (d**2) + (2*center_y*b) - (center_y**2) - (b**2)))/\
-        (1+m**2)
-    y = m*x + b
+    x = (center_x - (m * b) + (center_y * m) -
+         math.sqrt((d ** 2) * (m ** 2) - (m ** 2) * (center_x ** 2) + (2 * center_y * m * center_x) - (
+                 2 * m * b * center_x) + (d ** 2) + (2 * center_y * b) - (center_y ** 2) - (b ** 2))) / \
+        (1 + m ** 2)
+    y = m * x + b
     return x, y
 
 
 def create_street_segment(start_x, start_y, end_x, end_y, step):
-
     # y = mx + b
     if start_x == end_x:
         m = 0
@@ -79,15 +84,46 @@ def create_street_segment(start_x, start_y, end_x, end_y, step):
 def create_street_segments():
     pass
 
-def create_usable_area_polygons():
-    pass
+
+def calc_polygon_difference(polygon, subtracting_polygons):
+    diff_polygon = polygon
+
+    i = 0
+    for subtracting_polygon in subtracting_polygons:
+        diff_polygon = difference(diff_polygon, subtracting_polygon)
+        i += 1
+        # if i > 2:
+        #     break
+    return diff_polygon
 
 
-def check_polygons_usable_area():
+def remove_duplicate_points(multipolygon):
+    coordinates = multipolygon["geometry"]["coordinates"]
+    for polygon in coordinates:
+        rings_to_remove = []
+        for ring in polygon:
+            to_remove = []
+            for i in range(len(ring) - 1):
+                point = ring[i]
+                point2 = ring[i + 1]
+                if point == point2:
+                    to_remove.append(point2)
+            for point in to_remove:
+                ring.remove(point)
+            if len(ring) < 3:
+                rings_to_remove.append(ring)
+        for ring in rings_to_remove:
+            polygon.remove(ring)
+
+
+def create_walkable_area_polygons():
     cells = list(crowding_polygons.find({}))
-    #crowding_polygons.delete_many({})
+    # usable_areas.delete_many({})
+    usable_areas_list = []
     for cell in cells:
-        del cell["_id"]
+        if "_id" in cell:
+            del cell["_id"]
+    for cell in cells:
         intersecting_buildings = list(buildings.find({
             "geometry": {
                 "$geoIntersects": {
@@ -110,74 +146,76 @@ def check_polygons_usable_area():
             }
         }))
         unusable_polygons = intersecting_buildings + intersecting_water_bodies + intersecting_roads
+
+        unusable_polygon_union = union(FeatureCollection(unusable_polygons))
+        unusable_polygon_union = add_buffer_to_polygons([unusable_polygon_union], 0.10)
+        if unusable_polygon_union:
+            usable_area_polygon = difference(cell, unusable_polygon_union[0])
+        else:
+            usable_area_polygon = cell
+
+        usable_areas_list.append(usable_area_polygon)
+
+    # usable_areas.insert_many(usable_areas_list)
+    f = open("walkable_areas.json", "w")
+    f.write(json.dumps(usable_areas_list))
+
+def check_polygons_usable_area(cells):
+    #cells = list(crowding_polygons.find({}))
+    # crowding_polygons.delete_many({})
+    usable_areas.delete_many({})
+    usable_areas_list = []
+    for cell in cells:
+        if "_id" in cell:
+            del cell["_id"]
+        intersecting_buildings = list(buildings.find({
+            "geometry": {
+                "$geoIntersects": {
+                    "$geometry": cell["geometry"]
+                }
+            }
+        }))
+        intersecting_water_bodies = list(water.find({
+            "geometry": {
+                "$geoIntersects": {
+                    "$geometry": cell["geometry"]
+                }
+            }
+        }))
+        intersecting_roads = list(road_polygons.find({
+            "geometry": {
+                "$geoIntersects": {
+                    "$geometry": cell["geometry"]
+                }
+            }
+        }))
+        unusable_polygons = intersecting_buildings + intersecting_water_bodies + intersecting_roads
+
+
         unusable_area = 0
         for polygon in unusable_polygons:
             polygon_intersect = intersect([cell, polygon])
             intersect_area = area(polygon_intersect)
             unusable_area += intersect_area
+
         cell_area = area(cell)
         cell["properties"]["unusable_area"] = unusable_area
         cell["properties"]["usable_area"] = cell_area - unusable_area
         cell["properties"]["area"] = cell_area
 
-    f = open("crowding_polygons_with_usable_areas.json", "w")
-    #crowding_polygons.delete_many({})
-    f.write(json.dumps(cells))
+        usable_areas_list.append(cell)
+    return usable_areas_list
+    #usable_areas.insert_many(usable_areas_list)
+    # f = open("crowding_polygons_with_usable_areas.json", "w")
+    # f.write(json.dumps(cells))
 
-# ajuste ao algoritmo para pintar zonas pedonáveis
-def check_should_paint(feature):
-    intersecting_buildings = list(buildings.find({
-        "geometry": {
-            "$geoIntersects": {
-                "$geometry": feature["geometry"]
-            }
-        }
-    }))
-    intersecting_water_bodies = list(water.find({
-        "geometry": {
-            "$geoIntersects": {
-                "$geometry": feature["geometry"]
-            }
-        }
-    }))
-    intersecting_roads = list(road_polygons.find({
-        "geometry": {
-            "$geoIntersects": {
-                "$geometry": feature["geometry"]
-            }
-        }
-    }))
-    feature["properties"]["paint"] = (
-        len(intersecting_buildings) == 0
-        and len(intersecting_water_bodies) == 0
-        and len(intersecting_roads) == 0
-    )
-
-def check_all_usable_area():
-    features = list(osm.find({
-        "geometry.type": {"$regex": "(Multi)?Polygon"},
-    }))
-    res = []
-    for feature in features:
-        del feature["_id"]
-        if feature["geometry"]["type"] == "MultiPolygon":
-            for polygon_coordinates in feature["geometry"]["coordinates"]:
-                feat = {"type": "Feature", "geometry": {"type": "Polygon", "coordinates": polygon_coordinates}, "properties": deepcopy(feature["properties"])}
-                check_should_paint(feat)
-                res.append(feat)
-        else:
-            check_should_paint(feature)
-            res.append(feature)
-
-    with open("painted_polygons.json", "w", encoding="utf-8") as f:
-        json.dump(res, f)
 
 def create_crowding_polygons_from_sensors():
     crowding_polygons.delete_many({})
     sensors_list = list(sensors.find({}))
     polygons_list = []
     for sensor in sensors_list:
-        cc = circle(sensor, radius=0.390, steps=30)
+        cc = circle(sensor, radius=0.27475, steps=30)
         polygons_list.append(cc)
     crowding_polygons.insert_many(polygons_list)
 
@@ -202,11 +240,25 @@ def create_road_polygons():
     roads_df["geometry"] = roads_df["geometry"].apply(shape)
     roads_df = roads_df.set_geometry("geometry").set_crs("WGS84")
     roads_df["geometry"] = roads_df["geometry"].to_crs("EPSG:32633")
-    roads_df["geometry"] = roads_df["geometry"].buffer(roads_df["est_width"]/2, cap_style=2)
+    roads_df["geometry"] = roads_df["geometry"].buffer(roads_df["est_width"] / 2, cap_style=2)
     roads_df["geometry"] = roads_df["geometry"] = roads_df["geometry"].to_crs("WGS84")
     roads_json = roads_df.to_json()
     roads_dict = json.loads(roads_json)
     road_polygons.insert_many(roads_dict["features"])
+
+
+def add_buffer_to_polygons(polygons, buffer):
+    roads_df = gp.GeoDataFrame(polygons)
+    if "geometry" not in roads_df:
+        return None
+    roads_df["geometry"] = roads_df["geometry"].apply(shape)
+    roads_df = roads_df.set_geometry("geometry").set_crs("WGS84")
+    roads_df["geometry"] = roads_df["geometry"].to_crs("EPSG:32633")
+    roads_df["geometry"] = roads_df["geometry"].buffer(buffer, cap_style=2)
+    roads_df["geometry"] = roads_df["geometry"] = roads_df["geometry"].to_crs("WGS84")
+    roads_json = roads_df.to_json()
+    roads_dict = json.loads(roads_json)
+    return roads_dict["features"]
 
 
 def create_buildings_collection():
@@ -225,8 +277,9 @@ def create_buildings_collection():
                 if area(geojson) > max_area:
                     geometry["coordinates"] = polygon_coordinates
 
-    buildings.insert_many(buildings_list)
-    buildings.delete_many({"properties.layer": "-1"})
+    if len(buildings_list) > 0:
+        buildings.insert_many(buildings_list)
+        buildings.delete_many({"properties.layer": "-1"})
 
 
 def create_water_collection():
@@ -239,8 +292,28 @@ def create_water_collection():
         }
     ))
     water.delete_many({})
-    water.insert_many(water_list)
+    if len(water_list) > 0:
+        water.insert_many(water_list)
 
+
+def create_footways_collection():
+    footway_list = list(osm.find(
+        {
+            "properties.highway": {"$exists": True},
+            "geometry.type": "LineString",
+            "$or": [
+                {"properties.highway": "pedestrian"},
+                {"properties.highway": "footway"},
+                {"properties.highway": "steps"},
+            ],
+        }
+    ))
+    footways.delete_many({})
+    if len(footway_list) > 0:
+        footways.insert_many(footway_list)
+
+def create_tram_way_polygons():
+    pass
 
 def create_roads_collection():
     average_lane_width = 3  # METERS
@@ -261,13 +334,13 @@ def create_roads_collection():
     ))
 
     for road in roads_list:
-        
+
         num_lanes = int(road.get("properties").get("lanes", 1))
         est_width = num_lanes * average_lane_width
-        
+
         # ---- PARKING ---------
         has_parking_data = False
-
+        #TODO: Include tram lines
         if "parking:lane:left" in road:
             has_parking_data = True
             if road["parking:lane:left"] == "parallel":
@@ -298,6 +371,7 @@ def create_geo_indexes():
     water.create_index([("geometry", "2dsphere")])
     roads.create_index([("geometry", "2dsphere")])
     road_polygons.create_index([("geometry", "2dsphere")])
+    footways.create_index([("geometry", "2dsphere")])
 
 
 def create_osm_collections():
@@ -311,7 +385,6 @@ if __name__ == '__main__':
     create_osm_collections()
     create_geo_indexes()
     # create_crowding_polygons_from_sensors()
-    check_polygons_usable_area()
-    #check_all_usable_area()
-
-
+    # check_polygons_usable_area()
+    create_walkable_area_polygons()
+    # create_footways_collection()
